Adaptation to C++11: declarations and with arithmetic overflow
toto: warning / unnamed objects in construction & warnings from other software

+++ dbWave2/ExportDataDlg.h
@@ -28,7 +28,7 @@ public:
 	float	mm_timelast;
 	int		mm_firstchan;
 	int		mm_lastchan;
-	short	mm_binzero;
+	int		mm_binzero;
 	CAcqDataDoc* m_pDat;
 
 // Dialog Data
